chore: clear the outstanding lint findings
A u32::from on a value already u32, two imports left behind when the example
stopped halving frames, and a manual modulo test. None change behaviour;
they are cleared so the groups covering them can be denied without carrying
exceptions.

+++ crates/elementium-codec/src/vpx_codec.rs
@@ -263,7 +263,7 @@ impl Vp8Encoder {
 
         let flags = if self.force_keyframe {
             self.force_keyframe = false;
-            u32::from(VPX_EFLAG_FORCE_KF)
+            VPX_EFLAG_FORCE_KF
         } else {
             0
         };
